protocol: Make receivetime a float since it's derived from sv.time casted to float anyway
From bones_was_here

git-svn-id: svn://svn.icculus.org/twilight/trunk/darkplaces@13104 d7cf8633-e32d-0410-b094-e92efae38249

+++ protocol.h
@@ -389,7 +389,7 @@ typedef struct usercmd_s
 	int		cursor_entitynumber;
 
 	double time; // time the move is executed for (non-cl_movement is executed at receivetime)
-	double receivetime; // time the move was received at (used for ping)
+	float receivetime; // time the move was received at (used for ping)
 	int msec; // for predicted moves
 	int buttons;
 	int impulse;
